ptree: split out generic vs. Linux-specific tests
